Move setting prc_status in the session to an earlier step.
