+++ tests/activity/test_activity_validate_accepted_submission.py
@@ -188,60 +188,6 @@ class TestValidateAcceptedSubmission(unittest.TestCase):
     @patch.object(activity_module, "storage_context")
     @patch.object(activity_module, "get_session")
     @patch.object(cleaner, "storage_context")
-    @patch.object(cleaner, "is_prc")
-    def test_do_activity_is_prc_exception(
-        self,
-        fake_is_prc,
-        fake_cleaner_storage_context,
-        fake_session,
-        fake_storage_context,
-    ):
-        directory = TempDirectory()
-
-        # set a non-None session value to test string concatenation
-        self.session.store_value("cleaner_log", "")
-        fake_session.return_value = self.session
-        zip_file_path = os.path.join(
-            test_activity_data.ExpandArticle_files_source_folder,
-            "30-01-2019-RA-eLife-45644.zip",
-        )
-        resources = helpers.expanded_folder_bucket_resources(
-            directory,
-            test_activity_data.accepted_session_example.get("expanded_folder"),
-            zip_file_path,
-        )
-        fake_storage_context.return_value = FakeStorageContext(
-            directory.path, resources
-        )
-        fake_cleaner_storage_context.return_value = FakeStorageContext(
-            directory.path, resources
-        )
-        fake_is_prc.side_effect = ParseError()
-        # do the activity
-        result = self.activity.do_activity(input_data("30-01-2019-RA-eLife-45644.zip"))
-        self.assertEqual(result, True)
-        self.assertTrue(
-            self.activity.logger.logexception.startswith(
-                (
-                    "ValidateAcceptedSubmission, XML ParseError exception"
-                    " in cleaner.is_prc parsing XML file"
-                    " 30-01-2019-RA-eLife-45644.xml for file"
-                )
-            )
-        )
-        log_file_path = os.path.join(
-            self.activity.get_tmp_dir(), self.activity.activity_log_file
-        )
-        with open(log_file_path, "r", encoding="utf8") as open_file:
-            log_contents = open_file.read()
-        log_errors = [
-            line for line in log_contents.split("\n") if "ERROR elifecleaner:" in line
-        ]
-        self.assertEqual(len(log_errors), 1)
-
-    @patch.object(activity_module, "storage_context")
-    @patch.object(activity_module, "get_session")
-    @patch.object(cleaner, "storage_context")
     @patch.object(cleaner, "check_files")
     def test_do_activity_exception_unknown(
         self,
@@ -286,10 +232,8 @@ class TestValidateAcceptedSubmission(unittest.TestCase):
     @patch.object(activity_module, "get_session")
     @patch.object(cleaner, "storage_context")
     @patch.object(cleaner, "preprint_url")
-    @patch.object(cleaner, "is_prc")
     def test_do_activity_prc(
         self,
-        fake_is_prc,
         fake_preprint_url,
         fake_cleaner_storage_context,
         fake_session,
@@ -297,7 +241,7 @@ class TestValidateAcceptedSubmission(unittest.TestCase):
         fake_email_smtp_connect,
     ):
         directory = TempDirectory()
-
+        self.session.store_value("prc_status", True)
         # set a non-None session value to test string concatenation
         self.session.store_value("cleaner_log", "")
         fake_session.return_value = self.session
@@ -316,7 +260,6 @@ class TestValidateAcceptedSubmission(unittest.TestCase):
         fake_cleaner_storage_context.return_value = FakeStorageContext(
             directory.path, resources
         )
-        fake_is_prc.return_value = True
         fake_preprint_url.return_value = "https://doi.org/10.1101/2021.06.02.446694"
         fake_email_smtp_connect.return_value = FakeSMTPServer(directory.path)
         # do the activity
@@ -328,10 +271,8 @@ class TestValidateAcceptedSubmission(unittest.TestCase):
     @patch.object(activity_module, "get_session")
     @patch.object(cleaner, "storage_context")
     @patch.object(cleaner, "preprint_url")
-    @patch.object(cleaner, "is_prc")
     def test_do_activity_prc_sciety_failure(
         self,
-        fake_is_prc,
         fake_preprint_url,
         fake_cleaner_storage_context,
         fake_session,
@@ -339,7 +280,7 @@ class TestValidateAcceptedSubmission(unittest.TestCase):
         fake_email_smtp_connect,
     ):
         directory = TempDirectory()
-
+        self.session.store_value("prc_status", True)
         # set a non-None session value to test string concatenation
         self.session.store_value("cleaner_log", "")
         # empty the docmap_string value in the session
@@ -360,7 +301,6 @@ class TestValidateAcceptedSubmission(unittest.TestCase):
         fake_cleaner_storage_context.return_value = FakeStorageContext(
             directory.path, resources
         )
-        fake_is_prc.return_value = True
         preprint_doi = "10.1101/2021.06.02.999999"
         fake_preprint_url.return_value = "https://doi.org/%s" % preprint_doi
         fake_email_smtp_connect.return_value = FakeSMTPServer(directory.path)
@@ -380,10 +320,8 @@ class TestValidateAcceptedSubmission(unittest.TestCase):
     @patch.object(activity_module, "get_session")
     @patch.object(cleaner, "storage_context")
     @patch.object(cleaner, "preprint_url")
-    @patch.object(cleaner, "is_prc")
     def test_do_activity_prc_no_preprint_url(
         self,
-        fake_is_prc,
         fake_preprint_url,
         fake_cleaner_storage_context,
         fake_session,
@@ -410,7 +348,6 @@ class TestValidateAcceptedSubmission(unittest.TestCase):
         fake_cleaner_storage_context.return_value = FakeStorageContext(
             directory.path, resources
         )
-        fake_is_prc.return_value = True
         fake_preprint_url.return_value = None
         fake_email_smtp_connect.return_value = FakeSMTPServer(directory.path)
         # do the activity
@@ -429,7 +366,7 @@ class TestValidateAcceptedSubmission(unittest.TestCase):
         fake_storage_context,
     ):
         directory = TempDirectory()
-
+        self.session.store_value("prc_status", False)
         # set a non-None session value to test string concatenation
         self.session.store_value("cleaner_log", "")
         fake_session.return_value = self.session
